+++ idr_gallery/views.py
@@ -1,5 +1,7 @@
 
-from django.http import HttpResponseRedirect, HttpResponseBadRequest, Http404, JsonResponse
+
+from django.http import (HttpResponseRedirect, HttpResponseBadRequest,
+                         Http404, JsonResponse)
 from django.urls import reverse, NoReverseMatch
 import json
 import logging
@@ -8,7 +10,7 @@ import urllib
 from collections import defaultdict
 
 import omero
-from omero.rtypes import wrap, rlong, rstring
+from omero.rtypes import wrap, rlong
 from omeroweb.webclient.decorators import login_required, render_response
 from omeroweb.api.decorators import login_required as api_login_required
 from omeroweb.api.api_settings import API_MAX_LIMIT
@@ -20,7 +22,8 @@ from . import gallery_settings as settings
 from .data.background_images import IDR_IMAGES, TISSUE_IMAGES, CELL_IMAGES
 from .data.tabs import TABS
 from .version import VERSION
-from .utils import get_image_info, BIA_URL, parse_kvp_with_link, prefix_http, split_link
+from .utils import (get_image_info, BIA_URL,
+                    parse_kvp_with_link, prefix_http, split_link)
 
 try:
     from omero_mapr import mapr_settings
@@ -32,6 +35,7 @@ MAX_LIMIT = max(1, API_MAX_LIMIT)
 
 EMBL_EBI_PUBLIC_GLOBUS_ID = "47772002-3e5b-4fd3-b97c-18cee38d6df2"
 TABLE_NAMESPACE = "openmicroscopy.org/omero/bulk_annotations"
+BFF_URL = "https://bff.allencell.org/app"
 
 
 def redirect_with_params(viewname, **kwargs):
@@ -77,18 +81,22 @@ def index(request, super_category=None, conn=None, **kwargs):
                                                 value=keyval[1],
                                                 operator="contains")
             # handle e.g. ?query=Publication%20Authors:smith
-            # ?key=Publication+Authors&value=Smith&operator=contains&resource=container
+            # ?key=Publication+Authors&value=Smith&operator=
+            # contains&resource=container
             keyval = query.split(":", 1)
             if len(keyval) > 1 and len(keyval[1]) > 0:
-                # search for studies ("containers") and use "contains"
+                # search for studies ("containers") and
+                # use "contains" for "authors"
+                # but otherwise use "equals"
                 # to match previous behaviour
                 # NB: 'Name' needs to be 'name' for search-engine
                 key = "name" if keyval[0] == "Name" else keyval[0]
+                operator = "contains" if "author" in key.lower() else "equals"
                 return redirect_with_params('idr_gallery_search',
                                             key=key,
                                             value=keyval[1],
                                             resource="container",
-                                            operator="contains")
+                                            operator=operator)
             return HttpResponseBadRequest(
                 "Query should be ?query=key:value format")
     context = {'template': template}
@@ -104,6 +112,7 @@ def index(request, super_category=None, conn=None, **kwargs):
         context['category'] = super_category
     context["TABS"] = TABS
     context["VERSION"] = VERSION
+    context["BFF_URL"] = BFF_URL
 
     settings_ctx = get_settings_as_context()
     context = {**context, **settings_ctx}
@@ -126,9 +135,11 @@ def _escape_chars_like(query):
 @render_response()
 def study_page(request, idrid, format="html", conn=None, **kwargs):
 
-    if len(idrid) != 7 or not idrid.startswith("idr") or not idrid[3:].isdigit():
-        raise Http404("Invalid IDR ID. IDR IDs should be in the form idrXXXX")
-    
+    if (len(idrid) != 7 or not
+            idrid.startswith("idr") or not idrid[3:].isdigit()):
+        raise Http404("Invalid IDR ID. \
+                    IDR IDs should be in the form idrXXXX")
+
     # find Project(s) or Screen(s) with this IDRID
     # query_service = conn.getQueryService()
     # params = omero.sys.ParametersI()
@@ -139,13 +150,14 @@ def study_page(request, idrid, format="html", conn=None, **kwargs):
     # "like" search not working above. Just iterate and check names!
     objs = [p for p in conn.getObjects("Project") if p.name.startswith(idrid)]
     if len(objs) == 0:
-        objs = [s for s in conn.getObjects("Screen") if s.name.startswith(idrid)]
+        objs = [s for s in conn.getObjects("Screen")
+                if s.name.startswith(idrid)]
 
     if len(objs) == 0:
         raise Http404("No Project or Screen found for %s" % idrid)
-    
+
     # E.g."idr0098-huang-octmos", "idr0098-huang-octmos/experimentA", then "B"
-    objs.sort(key=lambda x: (len(x.name), x.id))
+    objs.sort(key=lambda x: (len(x.name), x.name))
 
     # Use first object for KVPs
     pids = None
@@ -154,8 +166,9 @@ def study_page(request, idrid, format="html", conn=None, **kwargs):
         pids = [objs[0].id]
     else:
         sids = [objs[0].id]
-    anns, experimenters = marshal_annotations(conn, project_ids=pids, screen_ids=sids,
-                                              ann_type="map", ns="idr.openmicroscopy.org/study/info")
+    anns, experimenters = (marshal_annotations(conn, project_ids=pids,
+                           screen_ids=sids, ann_type="map",
+                           ns="idr.openmicroscopy.org/study/info"))
     kvps = defaultdict(list)
     for ann in anns:
         for kvp in ann["values"]:
@@ -163,28 +176,55 @@ def study_page(request, idrid, format="html", conn=None, **kwargs):
 
     # Choose Study Title first, then Publication Title
     title_values = kvps.get("Study Title", kvps.get("Publication Title"))
+    if settings.BASE_URL is not None:
+        base_url = settings.BASE_URL
+    else:
+        base_url = request.build_absolute_uri(reverse('index'))
+    bff_url = f"{base_url}searchengine/api/v1/resources/container_bff_data/"
     containers = []
     for obj in objs:
-        desc = obj.description
+        desc = obj.description if obj.description is not None else ""
         for token in ["Screen", "Project", "Experiment", "Study"]:
             if f"{token} Description" in desc:
                 desc = desc.split(f"{token} Description", 1)[1].strip()
+        otype = "project" if obj.OMERO_CLASS == "Project" else "screen"
+        study_bff_url = f"{bff_url}?container_name={obj.name}\
+                        &container_type={otype}"
+        # https://idr-testing.openmicroscopy.org/searchengine//api/v1/resources/
+        # container_bff_data/?container_name=idr0164-alzubi-hdbr%2FexperimentA&
+        # container_type=project
+
         containers.append({
             "id": obj.id,
             "name": obj.name,
+            "short_name": obj.name.split("/")[-1],
             "description": desc,
             "type": "Project" if obj.OMERO_CLASS == "Project" else "Screen",
             "kvps": kvps,
+            "csv_download": f"{study_bff_url}&file_type=csv",
+            "parquet_download": f"{study_bff_url}&file_type=parquet",
+            "bff_url_csv": get_bff_url(request, study_bff_url,
+                                       f"{obj.name}.csv", ext="csv"),
+            "bff_url_parquet": get_bff_url(request, study_bff_url,
+                                           f"{obj.name}.parquet",
+                                           ext="parquet"),
+            "empty_study_container": ("experiment" not in obj.name
+                                      and "screen" not in obj.name)
         })
 
     img_objects = []
     for obj in containers:
-        img_objects.extend(_get_study_images(conn, obj["type"], obj["id"], tag_text="Study Example Image"))
+        img_objects.extend(_get_study_images
+                           (conn, obj["type"], obj["id"],
+                            tag_text="Study Example Image"))
 
     if len(img_objects) == 0:
-        # None found with Tag - just load untagged image
-        img_objects = _get_study_images(conn, obj["type"], obj["id"])
-    images = [{"id": o.id.val, "name": o.name.val} for o in img_objects]
+        for obj in containers:
+            # None found with Tag - just load untagged image
+            img_objects.extend(_get_study_images
+                               (conn, obj["type"], obj["id"]))
+    images = [{"id": o.id.val, "name": o.name.val}
+              for o in img_objects]
 
     # Use first image to get download & path info...
     img_info = get_image_info(conn, images[0]["id"])
@@ -195,19 +235,24 @@ def study_page(request, idrid, format="html", conn=None, **kwargs):
     bia_ngff_id = None
     idrid_name = containers[0]["name"].split("/")[0]
     if data_location == "IDR" or data_location == "Github":
-        # then link to Download e.g. https://ftp.ebi.ac.uk/pub/databases/IDR/idr0002-heriche-condensation/
+        # then link to Download e.g.
+        # https://ftp.ebi.ac.uk/pub/databases/IDR/idr0002-heriche-condensation/
         # e.g. idr0002-heriche-condensation
         download_url = f"https://ftp.ebi.ac.uk/pub/databases/IDR/{idrid_name}"
 
     if data_location == "Embassy_S3":
-        # "mkngff" data is at https://uk1s3.embassy.ebi.ac.uk/bia-integrator-data/pages/idr_ngff_data.html
+        # "mkngff" data is at
+        # https://uk1s3.embassy.ebi.ac.uk/bia-integrator-data/
+        # pages/idr_ngff_data.html
         bia_ngff_id = img_path.split(BIA_URL, 1)[-1].split("/", 1)[0]
 
-    KNOWN_KEYS = ["Publication Authors", "Study Title", "Publication Title", "Publication DOI", "Data DOI", "License", 
-                  "PubMed ID", "PMC ID", "Release Date", "External URL", "Annotation File", "BioStudies Accession"]
+    known_keys = ["Publication Authors", "Study Title", "Publication Title",
+                  "Publication DOI", "Data DOI", "License",
+                  "PubMed ID", "PMC ID", "Release Date", "External URL",
+                  "Annotation File", "BioStudies Accession"]
     other_kvps = []
     for k, v in kvps.items():
-        if k in KNOWN_KEYS:
+        if k in known_keys:
             continue
         for value in v:
             other_kvps.append([k, value])
@@ -220,6 +265,7 @@ def study_page(request, idrid, format="html", conn=None, **kwargs):
 
     context = {
         "template": "idr_gallery/idr_study.html",
+        "base_url": base_url,
         "globus_origin_id": EMBL_EBI_PUBLIC_GLOBUS_ID,
         "idr_id": idrid,
         "idrid_name": idrid_name,
@@ -237,9 +283,11 @@ def study_page(request, idrid, format="html", conn=None, **kwargs):
         "license": parse_kvp_with_link("License", kvps),
         "pubmed_id": parse_kvp_with_link("PubMed ID", kvps),
         "pmc_id": parse_kvp_with_link("PMC ID", kvps),
-        "release_date": kvps.get("Release Date")[0] if "Release Date" in kvps else None,
-        "external_urls": [prefix_http(url) for url in kvps.get("External URL", [])],
-        "annotation_files": [split_link(link) for link in kvps.get("Annotation File", [])],
+        "release_date": kvps.get("Release Date")[0] if "Release Date" in kvps else None, # noqa
+        "external_urls": [prefix_http(url)
+                          for url in kvps.get("External URL", [])],
+        "annotation_files": [split_link(link)
+                             for link in kvps.get("Annotation File", [])],
         "bia_accession": parse_kvp_with_link("BioStudies Accession", kvps),
         "other_kvps": other_kvps,
         "jsonld": json.dumps(jsonld, indent=2),
@@ -396,10 +444,13 @@ def _get_study_images(conn, obj_type, obj_id, limit=1,
     params.theFilter = omero.sys.Filter()
     params.theFilter.limit = wrap(limit)
     params.theFilter.offset = wrap(offset)
+    fetch_anns = ""
     and_text_value = ""
     if tag_text is not None:
         params.addString("tag_text", tag_text)
         and_text_value = " and annotation.textValue = :tag_text"
+        fetch_anns = " left outer join i.annotationLinks as al"\
+                     " join al.child as annotation"
 
     if obj_type.lower() == "project":
         query = "select i from Image as i"\
@@ -407,8 +458,7 @@ def _get_study_images(conn, obj_type, obj_id, limit=1,
                 " join dl.parent as dataset"\
                 " left outer join dataset.projectLinks"\
                 " as pl join pl.parent as project"\
-                " left outer join i.annotationLinks as al"\
-                " join al.child as annotation"\
+                + fetch_anns + \
                 " where project.id = :id%s" % and_text_value
 
     elif obj_type.lower() == "screen":
@@ -418,8 +468,7 @@ def _get_study_images(conn, obj_type, obj_id, limit=1,
                  " join well.plate as pt"
                  " left outer join pt.screenLinks as sl"
                  " join sl.parent as screen"
-                 " left outer join i.annotationLinks as al"
-                 " join al.child as annotation"
+                 + fetch_anns +
                  " where screen.id = :id%s"
                  " order by well.column, well.row" % and_text_value)
 
@@ -468,6 +517,7 @@ def api_thumbnails(request, conn=None, **kwargs):
         except KeyError:
             logger.error("Thumbnail not available. (img id: %d)" % i)
     return rv
+
 
 @login_required()
 @render_response()
@@ -541,3 +591,50 @@ def image_viewer(request, iid, conn=None, **kwargs):
         }
 
     return rsp_json
+
+
+def get_bff_url(request, data_url, fname, ext="csv"):
+    """
+    We build config into query params for the BFF app
+    """
+    data_url = f"{data_url}&file_type={ext}"
+    data_url = request.build_absolute_uri(data_url)
+    source = {
+        "uri": data_url,
+        "type": ext,
+        "name": fname,
+    }
+    s = urllib.parse.quote(json.dumps(source))
+    bff_url = f"{BFF_URL}?source={s}"
+    return bff_url
+
+
+@render_response()
+def download_urls(request, conn=None, **kwargs):
+    """
+    Return a page with download URLs for all studies.
+    """
+    context = {
+        "template": "idr_gallery/download_urls.html",
+        "VERSION": VERSION,
+    }
+    # settings_ctx = get_settings_as_context()
+    # context = {**context, **settings_ctx}
+    return context
+
+
+def link_check(request):
+    """
+    API endpoint to check if a URL is valid, used by download_urls page
+    """
+    url = request.GET.get("url")
+    if url is None:
+        return HttpResponseBadRequest("Missing 'url' parameter")
+    try:
+        response = requests.head(url, allow_redirects=True, timeout=5)
+        is_valid = response.ok
+    except requests.RequestException:
+        is_valid = False
+    status_code = response.status_code if is_valid else None
+    return JsonResponse({"url": url, "is_valid": is_valid,
+                         "status_code": status_code})